+++ test/dummy/app/assets/javascripts/application.js
@@ -7,3 +7,4 @@
 
 //= require folio/dropzone/dropzone
 //= require folio/image/image
+//= require folio/leads/form/form