+++ test/dummy/app/controllers/home_controller.rb
@@ -6,4 +6,7 @@ class HomeController < ApplicationController
 
   def dropzone
   end
+
+  def lead_form
+  end
 end